Crude ver. of Cast Cone Calculator

+++ eucVector.h
@@ -7,17 +7,17 @@
 class EucVector 
 {
 private:
-    float direction = 0.0f;
-    float magnitude = 0.0f;
-    std::vector<int> values = {0, 0};
+    double direction = 0.0f;
+    double magnitude = 0.0f;
+    std::vector<double> values = {0, 0};
 
 public: 
-    EucVector(int x1, int y1);
+    EucVector(double x1, double y1);
 
     // Getters and Setters
-    float GetMagnitude(); // Formula: sqrt((y2-y1)^2 + (x2-x1)^2)
+    double GetMagnitude(); // Formula: sqrt((y2-y1)^2 + (x2-x1)^2)
 
-    float GetDirection(); // Formula: tan((y2-y1)/(x2-x1))
+    double GetDirection(); // Formula: tan((y2-y1)/(x2-x1))
 
-    std::vector<int> GetValues();
+    std::vector<double> GetValues();
 };
